AsyncHttp\Client – Don't combine gzip encoding with the HTTP Range header.
When we're requesting a byte range AND gzipped transfer encoding,
our intention is to get compressed bytes 0-X of the original file.

However, some servers will compress the file first, and then return
the compressed bytes 0-X. The result is both unpredictable and impossible
to decompress.

This commit also makes other cosmetic changes – exposes http response as
an integer, and ensures the request headers are lowercase